move everything under src/demux
